r3729/sofa-dev : FIX: compilation with CUDA 2.1 and colladadom on Windows
 //partial sync (29/35 changes)//

+++ modules/sofa/gpu/cuda/CudaTetrahedronTLEDForceField.cpp
@@ -457,7 +457,7 @@ void CudaTetrahedronTLEDForceField::ComputeDhDxTetra(const Element& e, const Vec
             DhDx[j][k] = 0;
             for (int m = 0; m < 3; m++)
             {
-                DhDx[j][k] += DhDr[j][m]*invJ[k][m];
+                DhDx[j][k] += (float)(DhDr[j][m]*invJ[k][m]);
             }
         }
     }
